requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.8
botocore==1.35.8
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.1.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.2
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0 
s3transfer==0.10.2
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchtower==3.3.1
Werkzeug==3.0.3
WTForms==3.1.2
xero-python==1.12.0
